Add standalone bundle that runs bundle installation tests
The test bundle can be found at
testsuite/performance/target/test-libs/standalone-bundles/bundle-perf-test.jar
this bundle does not have a dependency on shrinkwrap or arquillian.
However the test bundle depends on the fact that the bundles being
installed are created beforehand. This is done with the
 org.jboss.osgi.test.performance.bundle.tool.BundleCreator tool
which can be run from the command line.
The maven build runs this tool automatically during the test-compile
phase for 25 bundles. If a different population of test bundles is
required the BundleCreator tool must be run again with the desired
number of bundles as argument.

The bundle-perf-test.jar bundle needs to be configured with the
following system properties:
    org.jboss.osgi.test.performance.bundles.dir
  The directory wher the test bundles are (i.e. where BundleCreator
has created them). Default '.'

    org.jboss.osgi.test.performance.bundle.size
  The population size (i.e. the number for test bundles being
installed). Default 25.